+++ app/code/community/LimeSoda/EnvironmentConfiguration/Block/Adminhtml/Overview.php
@@ -12,6 +12,37 @@ class LimeSoda_EnvironmentConfiguration_Block_Adminhtml_Overview extends Mage_Ad
     }
 
     /**
+     * @copyright Tatu Ulmanen (http://stackoverflow.com/a/2915920)
+     * @param array|null $tree
+     * @return string
+     */
+    public function getEnvironmentTreeHtml(array $tree = null)
+    {
+        $result = '';
+        if(!is_null($tree) && count($tree) > 0) {
+            $result .= '<ul class="environment-list">';
+            foreach($tree as $node) {
+                $url = $this->getUrl('*/*/*', array('env' => $node['name']));
+                $name = $this->escapeHtml($node['name']);
+                $result .= '<li class="environment-list-node">';
+                if ($this->hasEnvironment() && $this->getEnvironment() == $node['name']) {
+                    $result .= '<span class="environment-list-selected">' . $name . '</span>' .
+                                ' [ <a href="' . $this->getUrl('*/*/*') . '">' . $this->__('Unselect') . '</a> ]';
+                } else {
+                    $result .= '<a href="' . $url . '">' . $name . '</a>';
+                }
+                if (Mage::helper('limesoda_environmentconfiguration/current')->getEnvironmentName() == $node['name']) {
+                    $result .= ' [ ' . $this->__('Current environment') . ' ]';
+                }
+                $result .= $this->getEnvironmentTreeHtml($node['children']);
+                $result .= '</li>';
+            }
+            $result .= '</ul>';
+        }
+        return $result;
+    }
+
+    /**
      * @param string $command
      * @return string
      */