+++ CHANGELOG.md
@@ -7,8 +7,10 @@ Changelog
 Added / Improved:
 
 * Added node <system_configuration> which allows specifying system configuration settings in a more readable manner.
-* Added backend overview which shows the configured environments and the commands to be executed in a selected
-  environment. Missing variables are highlighted.
+* Added backend overview which:
+  * shows the environments as a hierarchy
+  * shows the commands to be executed in a selected environment
+  * highlights missing variables in the selected environment
 * Added new command stages 'pre_configure' and 'post_configure'.
 * Added n98-magerun command db:truncate.
 * Added n98-magerun command ls:env:reset.